Add function to get organism name from XML file directly if not supplied by user
Previously if user wanted to use a local XML file, they still had to specify an
organism abbreviation on the command line. Since the organism name is
stored in the XML file, we might as well check for it in cases where the org
name is omitted from the script args.

Minor reformat of usage string grammar.