Add malformed file check for edge cases where file cannot be found on server
I haven't worked out all the details, but there can be an issue when retrieving files
kept in tape archives on JGI's servers. For example, running on the query
"Metazome" produces a long list of available data, much of which is listed as being
stored in tape archives on the Metazome website. The cURL request doesnt seem
to properly trigger whatever backend voodoo has to happen for the file to become
available, and so instead the file that is downloaded is an XML error report. One
roundabout solution to this  is to request the file through the web interface, and
then cancel the download. This seems to be enough (in the limited cases I've tested)
to rejigger the backend and allow jgi-query to grab the correct file upon resubmitting
the same request. YMMV, of course, but the added code will at least alert the user to
the problem, and will avoid decompression errors.